moves auth components into components folder and updates routing
